Fixed problem related to product entity updates modified mapping for description so it can fit a long string of characters

+++ src/main/java/com/productosisis/models/service/ProductService.java
@@ -12,8 +12,8 @@ import java.util.Optional;
 import com.productosisis.models.entity.Product;
 
 public interface ProductService {
-	public Optional<Product> findById(int id);
+	public Optional<Product> findById(Long id);
 	public List<Product> findAll();
 	public Product save(Product p);
-	public void deleteById(int id);
+	public void deleteById(Long id);
 }
